+++ src/constraint_system/boolean.rs
@@ -53,7 +53,7 @@ mod tests {
     fn test_correct_bool_gate() {
         let res = gadget_tester(
             |composer| {
-                let zero = composer.add_input(BlsScalar::zero());
+                let zero = composer.zero_var();
                 let one = composer.add_input(BlsScalar::one());
 
                 composer.boolean_gate(zero);